Add type for nodes on the list
Since the idea is to test for memory leaks after using pointers, I chose
to use pointers for the data in the list as well. This allows me also
reuse the code for lists of any data types. Besides which, a list should
have no concern with what the data is, only the list itself.

The flip-side of using pointers to the data is that I'm going to need to
create helper routines to handle the data through its pointers, since
list, which has no clue about it, cannot handle the data.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>