Add pictographs for clarity
Using ASCII chars to display the list concept as it grows. It was a good
idea, but takes a lot of space. Not sure it's going to be worth it.

Might be a better idea to create a separate markdown file, but then it's
no longer in the code it's planning for or explaining.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>