Add display procedures
Only created procedures for intergers and floats. Strings, arrays,
assorted other numbers, and complex structures all can be displayed
using a purpose built procedure. From the standpoint of the list
operations, however, it is all the same operationally. Made two so that
I can test what happens if I create data of one type and, as if in
error, pass the other type's display procedure to walkNodes.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>