Change nomenclature for pointers
Using the word "node" in variable names could be misleading, as they do
not point to the node their name suggests. Rather they will point to the
node which points to the node suggested by the name. Hence, I switched
all the '*_node' variables to '*_ptr' names. I can still use "node" for
occasions where it is appropriate, and likely to be understood for what
it realy is.

Included at the top of the file a comment explaining the use of pointer
within the program.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>