Add procedure to add nodes to the list
Of note, again, is that the list procedures have no concern with the
type of data in the list. In that light, the "data" added to the list is
only a pointer to that data, and that pointer has to be created external
to the procedure and passed in by the caller.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>