Remove pictograph comments
As feared, the pictographs were nice, but big. With only two of the
procedures coded for the linked list, they consumed 63% of the lines of
code. I have a picture in my mind, which ASCII can't display anyway, and
anyone else will either form their own mental map of the process, or
they have already done this and don't need one anyway.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>