Add procedure to add a node to the head of the list
The procedure prependNode adds a node to the head of the list, while the
old addNode procedure, renamed to appendNode, adds the node to the tail
of the list.

The addNode procedure was changed to appendNode to reflect what it does,
and to remove ambiguity. Adding a node could be done anywhere in the
list: head, tail, or anywhere in between.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>