Add procedures to insert a node into the list
Two version are given, insertNodeAfter, and insertNodeBefore. Both will
insert a node, with the supplied data pointer. The "before" or "after"
is relative to curr_ptr->next, in keeping with the stated use of *_ptr
variables. If curr_ptr is the same as head_ptr, "before" would be the
same effect as prepending the node, while "after" would make the new
node the second one in the list.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>