Add procedure to find a node in the list
I need to be able to find specific nodes in order to use them as places
to insert into the list, or for what to delete/remove from the list. As
usual for this project, the pointer returned is once removed in that the
returned pointer points to the node that points to the node with the
match in it.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>