Add comparison functions for integers and floats
Need some functions for testing the values in the findNode function. As
only integers and floats are being used for the display, I only need the
same for tests. Real world tests could be built that would check the
contents of a few fields in a complete record, for example.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>