Add procedure to clear the list
The procedure will clear out the list. It will not, however, erase the
information in the data field. Clearing the data from the list is the
responsibility of the owner of that data, who should know how to do so.

Allowing for the clearing of the data means that the walkList needs to
be repurposed to do any action the caller wishes, not just display the
data. No change to the procedure itself, just a variable name change to
reflect its purpose.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>