Change variable names to be more uniform
Wanted to elimiate the use of curr_ptr and temp or temp_ptr, in favor of
index_ptr. The curr_ptr might be implemented as a working variable for
the list, and didn't want it to be ambiguously present elsewhere. Since
the pointer is usually used while indexing through the list, the name
made sense.

A few minor cleanups were done at the same time, just as a housekeeping
measure.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>