Update insertNodeAfter for empty list
I realized that trying to insert a node after an empty list was a) not
really what someone should do, and b) was possible, but not handled by
the program. If someone tries to do that, I will just make that the new
list.

There is no programatic access to the last node address, a search for it
returns a ListNode* to the node that points to the last node, the second
last node, so trying to call insertNodeAfter with the actual last node
as the pointer value requires that the caller by trying to bypass the
in-built process. Still, if attempted, the new node is just appended as
if nothing had been wrong.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>