Add sanity checks to the list procedures
Most of the sanity checks are for passing in NULL when it should be an
active pointer. An empty list, or a NULL valued next pointer are also
common checks.

In some cases there is an early exit when there is nothing to do; either
and empty list to walk, or no procedure passed in to perform, etc.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>