Add check NULL target in findNode
No point doing a function call to test equality of two nodes when the
target is a NULL pointer. Two options exist. Either NULL is a match for
NULL, or NULL matches nothing, including NULL. Since storing NULL in the
list is valid, there needs to be a way to find such nodes on the list.
Therefore, I used the option that NULL == NULL.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>