Modify sanity check in removeNode and deleteNode
Passing in a NULL pointer is a sanity check failure. Passing in a
pointer that has no node attached, however, is "legal" even if wastefull
as there is nothing to delete/remove. In production code it should
probably be a silent drop, but it can stay here with its own message for
the challenge.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>