Change name of variable in findNode
Changed match_proc to match_func since many people consider procedures
as routines that do not return values and functions as routines that do
return values. In findNode, the linked routine is expected to return a
value (Boolean), while in walkList the routine is not expected to return
anything (void).

Just a nod to some convention somewhere.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>