Add optional procedure call to free node data
Added clear_proc pointer to the parameters for clearList and deleteNode
so that any needed cleanup can be done as the node(s) are being freed.
This saves the hassle of having to make a separate trip through the list
to free the memory, or do other cleanup, as part of clearing the list.

The choice to add it to deleteNode was just because it seemed best to
make clearList, which deletes all nodes, and deleteNode, which only
deletes one node, to look and act the same in their handling of the list
nodes. Not adding it to removeNode is because removeNode returns the
node data, and the caller is expected to handle it anyway they wish,
including deleting it, but not limited to that.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>