Factor routines into related files
Reduce the size of the main file by converting the code into realted
files to include. The bulk of the code is still in one file anyway, but
the main file is now "clean."

For data types other than float and int, a companion file to floathelp
and inthelp will need to be made, and included in the main file. All the
routines in inthelp need to be replicated for the new data type.

During the split I also did some clean up of the code, expanding some of
the comments, making conditionals follow a common pattern (literals and
constants on the left).

The main file is renamed to challenge-main.c

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>