Build all the tests into one pass
All the previous test. This time, all run in the same program, one after
the other.

List size can be increased as needed to make the process take longer.

The random number test and the remote test both can be adjusted for size.

Absent finding error or bugs, this is now ready for profiling in
valgrind to check for memory leaks. The whole point to this exercise.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>