Prep for Valgrind testing
All the written tests have passed, by my inspection, anyway.

Comment out all the tests but the last one. Use that one for profiling
the system.

A reasonably high setting for the count should exercise the system, and
give it a chance to leak. At the same time, not sure I want to wait for
a couple hours to see what it does either.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>