Clean up minor oversights
In the testing code there were a few places where memory was not freed.
This is not really a problem since, 1) they are one-shot deals and then
they exit, and 2) they are not production code. Since they exit after a
single pass, all memory is reclaimed by the system anyway, so there is
not really a memory "leak". They are also not part of the challenge, and
I wouldn't expect them to be tested by valgrind. I certainly had not
planned to do so once I install it. On the flip side, good habits are
just good habits. If I see a problem with my code, I fix it. That it is
of no import is of no consequence.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>