Update README.md with Valgrind info
Valgrind is now installed and ready to use. The warning about speed and
memory use has convinced me to downgrade the size of the test. The test
is only going to be for 75k numbers rather than the 234567 I originally
set in the file. The _remote.c file seems like the best choice to use in
testing. It uses dynamic memory allocation for the list and for the data
on the list. The head_ptr for the list is independent of the called
routines, unlike the other tests. Even at 75k, it ought to give a good
reading for Valgrind to capture.

Signed-off-by: Gypsy Spellweaver <gypsyspellweaver@gmail.com>